Use vite entrypoint for javascript
