+++ app/views/layouts/spotlight/base.html.erb
@@ -8,9 +8,8 @@
     <title><%= h(@page_title || application_name.to_s) %></title>
     <link href="<%= current_exhibit ? spotlight.opensearch_exhibit_catalog_url(current_exhibit, format: 'xml') : main_app.opensearch_catalog_url(format: 'xml') %>" title="<%= h(@page_title || application_name.to_str) %>" type="application/opensearchdescription+xml" rel="search"/>
     <%= favicon_link_tag 'favicon.ico' %>
-    <%# javascript_pack_tag 'application' %>
+    <%= vite_javascript_tag 'application' %>
     <link rel="stylesheet" href="https://use.typekit.net/yhr7zwc.css">
-    <%# stylesheet_pack_tag 'application' %>
     <% if current_exhibit %>
       <%= exhibit_stylesheet_link_tag "application" %>
     <% else %>
